Split UIKit+StackKit to UIView+StackKit folder files
