+++ Sources/StackKit/UIView+StackKit/StackKitCompatible.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+public class StackKitCompatible<Base> {
+    let view: Base
+    init(view: Base) {
+        self.view = view
+    }
+}
