+++ Sources/StackKit/UIView+StackKit/StackKitCompatibleProvider.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+public protocol StackKitCompatibleProvider {
+    associatedtype O
+    var stack: O { get }
+    static var stack: O { get }
+}
